refactor: reorganize code generation scripts by pipeline type
