+++ scripts/auto_generate.sh
@@ -1,2 +1,5 @@
-uv run scripts/insert_methods.py
+uv run scripts/insert_pipeline_methods.py
+uv run scripts/insert_pipeline_lazy_methods.py
+uv run scripts/insert_group_by_methods.py
+uv run scripts/insert_group_by_lazy_methods.py
 uv format
